New implementation of isAlive in Chord.
isAlive now uses process registration on nodes to identify whether
remote nodesare still alive. TODO: there are currently 3 functions in
Chord that use isAlive. The use of isAlive is a bottleneck depending
on the frequency of pings. A better way to handle this would be to
spawnLocal a function that looks up the pid of the remote
"ping-server" registered process, and monitors it. We then call
`expect` and if the remote node fails, a ProcessMonitorNotification
will be received, which could be used to remove the node from finger
entries.